feat: pole of inaccessibility
