+++ src/index.js
@@ -2,7 +2,7 @@ import Map from './Map'
 import supportedLayers from './layers/layerTypes'
 import supportedControls from './controls/controlTypes'
 import getEarthEngineApi from './utils/eeapi'
-import { poleOfInaccessibility } from './utils/geometry'
+import { getLabelPosition } from './utils/labels'
 
 /**
  *  Wrapper around Mapbox GL JS for DHIS2 Maps
@@ -14,6 +14,6 @@ export const controlTypes = Object.keys(supportedControls)
 
 export const loadEarthEngineApi = getEarthEngineApi
 
-export const poleOfInaccessibility
+export const poleOfInaccessibility = getLabelPosition
 
 export default Map
