+++ src/utils/labels.js
@@ -1,4 +1,6 @@
-import { featureCollection, poleOfInaccessibility } from './geometry'
+import area from '@turf/area'
+import polylabel from 'polylabel'
+import { featureCollection } from './geometry'
 
 // Default fonts
 const fonts = {
@@ -19,7 +21,7 @@ export const labelSource = (features, { fontSize }, isBoundary) => ({
             type: 'Feature',
             geometry: {
                 type: 'Point',
-                coordinates: poleOfInaccessibility(geometry),
+                coordinates: getLabelPosition(geometry),
             },
             properties: {
                 name: properties.name,
@@ -84,4 +86,25 @@ export const labelLayer = ({
             'text-color': color ? color : ['get', 'color'],
         },
     }
+}
+
+export const getLabelPosition = ({ type, coordinates }) => {
+    if (type === 'Point') {
+        return coordinates
+    }
+
+    let polygon = coordinates
+
+    if (type === 'MultiPolygon') {
+        const areas = coordinates.map(coords =>
+            area({
+                type: 'Polygon',
+                coordinates: coords,
+            })
+        )
+        const maxIndex = areas.indexOf(Math.max.apply(null, areas))
+        polygon = coordinates[maxIndex]
+    }
+
+    return polylabel(polygon, 0.1)
 }
